Refactor log parsing script for improved clarity and efficiency
