+++ package.json
@@ -22,7 +22,7 @@
     "lint": "eslint .",
     "medium": "yarn test suites/medium",
     "monitor:dev": "httpstat https://grpc.dev.xmtp.network:443",
-    "parse-ansi": "tsx scripts/parse-ansi-logs.ts",
+    "parse-ansi": "tsx scripts/parse-ansi.ts",
     "railway": "railway run yarn test ",
     "record": "npx playwright codegen 'https://xmtp.chat/'",
     "script": "yarn cli script",
